+++ src/api/services/useOtherUser.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getUserByRandomId } from "@/api/services/getUserByRandomId";
+import { getUserByRandomId } from "@/api/users";
 import { User } from "@/api/users";
 
 const useOtherUser = (randomId: number) => {
